refactor ui and features

+++ src/components/CartItem/CartItem.js
@@ -1,7 +1,7 @@
+import "./CartItem.scss";
 import { useDispatch } from "react-redux";
 import { removeCart, decreaseCart, increaseCart } from "../../redux/cartSlice";
 import { numberFormat } from "../../utils/numberFormat";
-import "./CartItem.scss";
 
 const CartItem = (props) => {
   const { item } = props;
@@ -11,31 +11,47 @@ const CartItem = (props) => {
     if (type === "plus") {
       dispatch(increaseCart(item));
     } else {
-      dispatch(decreaseCart(item));
+      if (item.cartQuantity === 1) {
+        return;
+      } else {
+        dispatch(decreaseCart(item));
+      }
     }
   };
 
   const handleRemoveCart = () => {
-    dispatch(removeCart(item));
+    const isCheck = window.confirm(
+      "Bạn muốn xoá sản phẩm này ra khỏi giỏ hàng?"
+    );
+    if (isCheck) {
+      dispatch(removeCart(item));
+    } else {
+      return;
+    }
   };
 
   const apiUrl = process.env.REACT_APP_BASE_IMG;
   return (
     <div className="cart__product__item">
-      <img
-        src={`${apiUrl}${item.image}`}
-        alt="product-img"
-        className="product__img"
-      />
-      <div className="product__name">
-        <div className="product__name-title">{item.name}</div>
+      <div className="product__img">
+        <img src={`${apiUrl}${item.image}`} alt="product-img" />
+      </div>
+
+      <div className="product__info">
+        <div className="product__name">{item.name}</div>
         <div className="product__qty">
           <div className="cart__item-qty">
             <div
               className="cart__item-qty-btn"
               onClick={() => updateQuantity("minus")}
             >
-              <i className="fa-solid fa-minus"></i>
+              <i
+                className={
+                  item.cartQuantity === 1
+                    ? "fa-solid fa-minus disable"
+                    : "fa-solid fa-minus"
+                }
+              ></i>
             </div>
             <div className="cart__item-qty-input">{item.cartQuantity}</div>
             <div
@@ -49,9 +65,7 @@ const CartItem = (props) => {
       </div>
       <div className="product__price">
         <div>{numberFormat(item.price)} đ</div>
-        <div onClick={() => handleRemoveCart()} style={{ display: "inline" }}>
-          <i className="fa-regular fa-trash-can"></i>
-        </div>
+        <div onClick={() => handleRemoveCart()}>Xóa</div>
       </div>
     </div>
   );
